Remove support for running interactive programs
This removes `mch_call_shell` code for feeding programs interactively. The
removed code was supporting interactive programs in the old GUI, but right now
we only have a terminal UI.

The code is currently safe to remove because interactive programs will just
simply take control of the terminal in cooked mode.